Split off form component of flasher
